--- Plugins/CustomPixelShader/CustomPixelShader.uplugin
@@ -1,21 +1,0 @@
-{
-    "PluginFileVersion" : 3,
-	
-	"FriendlyName" : "Custom Pixel Shader",
-	"Version" : 1,
-	"VersionName" : "1.0",
-	"CreatedBy" : "Temaran",
-	"CreatedByURL" : "temaran@gmail.com",
-	"EngineVersion" : "4.7.0",
-	"Description" : "An example stand alone pixel shader that can be run from a plugin",
-	"Category" : "Programming Rendering.Plugins",
-	
-    "Modules" :
-	[
-		{
-			"Name" : "CustomPixelShader",
-			"Type" : "Runtime",
-			"LoadingPhase" : "PostConfigInit"
-		}
-	]
-}
