+++ Plugins/ShaderCopyHelper/ShaderCopyHelper.uplugin
@@ -1,21 +1,23 @@
 {
-    "FileVersion" : 1,
-	
-	"FriendlyName" : "Shader Copy Helper",
-	"Version" : 1,
-	"VersionName" : "0.0.1",
-	"CreatedBy" : "Temaran",
-	"CreatedByURL" : "https://github.com/Temaran",
-	"EngineVersion" : "4.7.0",
-	"Description" : "This plugin will copy all files present in your Project/Shaders folder to your Engine/Shaders folder",
-	"Category" : "Rendering",
-	
-    "Modules" :
-	[
-		{
-			"Name" : "ShaderCopyHelper",
-			"Type" : "Developer",
-			"LoadingPhase" : "PostConfigInit"
-		}
-	]
+  "FileVersion": 3,
+  "Version": 1,
+  "VersionName": "1.0",
+  "FriendlyName": "ShaderEngineCopyHelper",
+  "Description": "This plugin will copy all files present in your Project/Shaders folder to your Engine/Shaders folder",
+  "Category": "Rendering",
+  "CreatedBy": "zameran [Originally by Temaran]",
+  "CreatedByURL": "zameran9@gmail.com [Originally by temaran@gmail.com]",
+  "DocsURL": "",
+  "MarketplaceURL": "",
+  "SupportURL": "",
+  "CanContainContent": true,
+  "IsBetaVersion": true,
+  "Installed": false,
+  "Modules": [
+    {
+      "Name": "ShaderCopyHelper",
+      "Type": "Developer",
+      "LoadingPhase": "PostConfigInit"
+    }
+  ]
 }
