+++ Plugins/ShaderEngine/Source/ShaderEngine/ShaderEngine.Build.cs
@@ -1,0 +1,60 @@
+using UnrealBuildTool;
+
+public class ShaderEngine : ModuleRules
+{
+	public ShaderEngine(TargetInfo Target)
+	{
+
+		PublicIncludePaths.AddRange(
+			new string[] 
+			{
+				"ShaderEngine/Public"
+			}
+		);
+
+
+		PrivateIncludePaths.AddRange(
+			new string[] 
+			{
+				"ShaderEngine/Private",
+			}
+		);
+
+		/*
+		"Core",
+		"CoreUObject",
+		"Engine",
+		"RenderCore",
+		"ShaderCore",
+		"RHI"
+		*/
+
+		PublicDependencyModuleNames.AddRange(
+			new string[]
+			{
+				"Core",
+				"RenderCore",
+				"ShaderCore",
+				"RHI"
+			}
+		);
+
+
+		PrivateDependencyModuleNames.AddRange(
+			new string[]
+			{
+				"CoreUObject",
+				"Engine",
+				"Slate",
+				"SlateCore",
+			}
+		);
+
+		DynamicallyLoadedModuleNames.AddRange(
+			new string[]
+			{
+
+			}
+		);
+	}
+}
