+++ Source/ShaderPluginDemo.Target.cs
@@ -1,25 +1,15 @@
-// Copyright 1998-2016 Epic Games, Inc. All Rights Reserved.
-
 using UnrealBuildTool;
+
 using System.Collections.Generic;
 
 public class ShaderPluginDemoTarget : TargetRules
 {
-	public ShaderPluginDemoTarget(TargetInfo Target)
-	{
-		Type = TargetType.Game;
-	}
-
-	//
-	// TargetRules interface.
-	//
-
-	public override void SetupBinaries(
-		TargetInfo Target,
-		ref List<UEBuildBinaryConfiguration> OutBuildBinaryConfigurations,
-		ref List<string> OutExtraModuleNames
-		)
-	{
-		OutExtraModuleNames.Add("ShaderPluginDemo");
-	}
+    public ShaderPluginDemoTarget(TargetInfo Target)
+    {
+        Type = TargetType.Game;
+    }
+    public override void SetupBinaries(TargetInfo Target, ref List<UEBuildBinaryConfiguration> OutBuildBinaryConfigurations, ref List<string> OutExtraModuleNames)
+    {
+        OutExtraModuleNames.Add("ShaderPluginDemo");
+    }
 }
