+++ Source/ShaderPluginDemo/ShaderPluginDemo.Build.cs
@@ -1,13 +1,11 @@
-// Copyright 1998-2016 Epic Games, Inc. All Rights Reserved.
-
 using UnrealBuildTool;
 
 public class ShaderPluginDemo : ModuleRules
 {
-	public ShaderPluginDemo(TargetInfo Target)
-	{
-		PublicDependencyModuleNames.AddRange(new string[] { "Core", "CoreUObject", "Engine", "InputCore", "RHI" });
+    public ShaderPluginDemo(TargetInfo Target)
+    {
+        PublicDependencyModuleNames.AddRange(new string[] { "Core", "CoreUObject", "Engine", "InputCore", "RHI" });
 
-        PublicDependencyModuleNames.AddRange( new string[] { "PixelShader", "ComputeShader", "CustomPixelShader" } );
+        PublicDependencyModuleNames.AddRange(new string[] { "ShaderEngine" });
     }
 }
