+++ Source/ShaderPluginDemo/ShaderPluginDemo.cpp
@@ -1,7 +1,3 @@
-// Copyright 1998-2016 Epic Games, Inc. All Rights Reserved.
-
 #include "ShaderPluginDemo.h"
 
-
-IMPLEMENT_PRIMARY_GAME_MODULE( FDefaultGameModuleImpl, ShaderPluginDemo, "ShaderPluginDemo" );
- 
+IMPLEMENT_PRIMARY_GAME_MODULE(FDefaultGameModuleImpl, ShaderPluginDemo, "ShaderPluginDemo");
