+++ Source/ShaderPluginDemo/ShaderPluginDemo.h
@@ -1,9 +1,3 @@
-// Copyright 1998-2016 Epic Games, Inc. All Rights Reserved.
-
-#ifndef __SHADERPLUGINDEMO_H__
-#define __SHADERPLUGINDEMO_H__
+#pragma once
 
 #include "Engine.h"
-
-
-#endif
