+++ Source/ShaderPluginDemo/ShaderPluginDemoCharacter.h
@@ -1,9 +1,11 @@
-// Copyright 1998-2016 Epic Games, Inc. All Rights Reserved.
 #pragma once
+
 #include "GameFramework/Character.h"
-#include "CustomPixelShader.h"
-// #include "PixelShaderUsageExample.h"
-#include "ComputeShaderUsageExample.h"
+
+#include "ShaderEngine/Public/ComputeShader/ComputeShaderUsageExample.h"
+#include "ShaderEngine/Public/CustomPixelShader/CustomPixelShaderUsageExample.h"
+#include "ShaderEngine/Public/PixelShader/PixelShaderUsageExample.h"
+
 #include "ShaderPluginDemoCharacter.generated.h"
 
 class UInputComponent;
@@ -14,7 +16,7 @@ class AShaderPluginDemoCharacter : public ACharacter
 	GENERATED_BODY()
 
 	/** Pawn mesh: 1st person view (arms; seen only by self) */
-	UPROPERTY(VisibleDefaultsOnly, Category=Mesh)
+	UPROPERTY(VisibleDefaultsOnly, Category = Mesh)
 	class USkeletalMeshComponent* Mesh1P;
 
 	/** Gun mesh: 1st person view (seen only by self) */
@@ -34,19 +36,19 @@ public:
 	virtual void BeginPlay();
 
 	/** Base turn rate, in deg/sec. Other scaling may affect final turn rate. */
-	UPROPERTY(VisibleAnywhere, BlueprintReadOnly, Category=Camera)
+	UPROPERTY(VisibleAnywhere, BlueprintReadOnly, Category = Camera)
 	float BaseTurnRate;
 
 	/** Base look up/down rate, in deg/sec. Other scaling may affect final rate. */
-	UPROPERTY(VisibleAnywhere, BlueprintReadOnly, Category=Camera)
+	UPROPERTY(VisibleAnywhere, BlueprintReadOnly, Category = Camera)
 	float BaseLookUpRate;
 
 	/** Gun muzzle's offset from the characters location */
-	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category=Gameplay)
+	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = Gameplay)
 	FVector GunOffset;
 
 	/** Sound to play each time we fire */
-	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category=Gameplay)
+	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = Gameplay)
 	class USoundBase* FireSound;
 
 	/** AnimMontage to play each time we fire */
@@ -54,7 +56,6 @@ public:
 	class UAnimMontage* FireAnimation;
 
 protected:
-	
 	/** Fires a projectile. */
 	void OnFire();
 
@@ -78,24 +79,23 @@ protected:
 
 	struct TouchData
 	{
-		TouchData() { bIsPressed = false;Location=FVector::ZeroVector;}
+		TouchData() { bIsPressed = false; Location = FVector::ZeroVector; }
 		bool bIsPressed;
 		ETouchIndex::Type FingerIndex;
 		FVector Location;
 		bool bMoved;
 	};
+
 	void BeginTouch(const ETouchIndex::Type FingerIndex, const FVector Location);
 	void EndTouch(const ETouchIndex::Type FingerIndex, const FVector Location);
 	void TouchUpdate(const ETouchIndex::Type FingerIndex, const FVector Location);
-	TouchData	TouchItem;
-	
-protected:
-	// APawn interface
-	virtual void SetupPlayerInputComponent(UInputComponent* InputComponent) override;
-	// End of APawn interface
+	TouchData TouchItem;
 
-	/* 
-	 * Configures input for touchscreen devices if there is a valid touch interface for doing so 
+protected:
+	virtual void SetupPlayerInputComponent(UInputComponent* InputComponent) override;
+
+	/*
+	 * Configures input for touchscreen devices if there is a valid touch interface for doing so
 	 *
 	 * @param	InputComponent	The input component pointer to bind controls to
 	 * @returns true if touch controls were enabled.
@@ -108,25 +108,18 @@ public:
 	/** Returns FirstPersonCameraComponent subobject **/
 	FORCEINLINE class UCameraComponent* GetFirstPersonCameraComponent() const { return FirstPersonCameraComponent; }
 
-
-
-
-
-	/************************************************************************/
-	/* Plugin Shader Demo variables!                                        */
-	/************************************************************************/
 public:
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = ShaderDemo)
-		FColor PixelShaderTopLeftColor;
+	FColor PixelShaderTopLeftColor;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = ShaderDemo)
-		float ComputeShaderSimulationSpeed;
+	float ComputeShaderSimulationSpeed;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = ShaderDemo)
-		UMaterialInterface* MaterialToApplyToClickedObject;
+	UMaterialInterface* MaterialToApplyToClickedObject;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = ShaderDemo)
-		UTextureRenderTarget2D* RenderTarget;
+	UTextureRenderTarget2D* RenderTarget;
 
 protected:
 	virtual void BeginDestroy() override;
@@ -134,7 +127,7 @@ protected:
 
 private:
 	// FPixelShaderUsageExample* PixelShading;
-	FCustomPixelShader* CustomPixelShading;
+	FCustomPixelShaderUsageExample* CustomPixelShading;
 	FComputeShaderUsageExample* ComputeShading;
 
 	float EndColorBuildup;
@@ -149,4 +142,3 @@ private:
 	void SavePixelShaderOutput();
 	void SaveComputeShaderOutput();
 };
-
