+++ Source/ShaderPluginDemo/ShaderPluginDemoHUD.cpp
@@ -1,35 +1,27 @@
-// Copyright 1998-2016 Epic Games, Inc. All Rights Reserved.
-
 #include "ShaderPluginDemo.h"
 #include "ShaderPluginDemoHUD.h"
+
 #include "Engine/Canvas.h"
 #include "TextureResource.h"
 #include "CanvasItem.h"
 
 AShaderPluginDemoHUD::AShaderPluginDemoHUD()
 {
-	// Set the crosshair texture
 	static ConstructorHelpers::FObjectFinder<UTexture2D> CrosshiarTexObj(TEXT("/Game/FirstPerson/Textures/FirstPersonCrosshair"));
+
 	CrosshairTex = CrosshiarTexObj.Object;
 }
-
 
 void AShaderPluginDemoHUD::DrawHUD()
 {
 	Super::DrawHUD();
 
-	// Draw very simple crosshair
-
-	// find center of the Canvas
 	const FVector2D Center(Canvas->ClipX * 0.5f, Canvas->ClipY * 0.5f);
+	const FVector2D CrosshairDrawPosition((Center.X - (CrosshairTex->GetSurfaceWidth() * 0.5)), (Center.Y - (CrosshairTex->GetSurfaceHeight() * 0.5f)));
 
-	// offset by half the texture's dimensions so that the center of the texture aligns with the center of the Canvas
-	const FVector2D CrosshairDrawPosition( (Center.X - (CrosshairTex->GetSurfaceWidth() * 0.5)),
-										   (Center.Y - (CrosshairTex->GetSurfaceHeight() * 0.5f)) );
+	FCanvasTileItem TileItem(CrosshairDrawPosition, CrosshairTex->Resource, FLinearColor::White);
 
-	// draw the crosshair
-	FCanvasTileItem TileItem( CrosshairDrawPosition, CrosshairTex->Resource, FLinearColor::White);
 	TileItem.BlendMode = SE_BLEND_Translucent;
-	Canvas->DrawItem( TileItem );
-}
 
+	Canvas->DrawItem(TileItem);
+}
