+++ Source/ShaderPluginDemo/ShaderPluginDemoHUD.h
@@ -1,6 +1,7 @@
-// Copyright 1998-2016 Epic Games, Inc. All Rights Reserved.
 #pragma once 
+
 #include "GameFramework/HUD.h"
+
 #include "ShaderPluginDemoHUD.generated.h"
 
 UCLASS()
@@ -17,6 +18,4 @@ public:
 private:
 	/** Crosshair asset pointer */
 	class UTexture2D* CrosshairTex;
-
 };
-
