+++ Source/ShaderPluginDemoEditor.Target.cs
@@ -1,25 +1,16 @@
-// Copyright 1998-2016 Epic Games, Inc. All Rights Reserved.
-
 using UnrealBuildTool;
+
 using System.Collections.Generic;
 
 public class ShaderPluginDemoEditorTarget : TargetRules
 {
-	public ShaderPluginDemoEditorTarget(TargetInfo Target)
-	{
-		Type = TargetType.Editor;
-	}
+    public ShaderPluginDemoEditorTarget(TargetInfo Target)
+    {
+        Type = TargetType.Editor;
+    }
 
-	//
-	// TargetRules interface.
-	//
-
-	public override void SetupBinaries(
-		TargetInfo Target,
-		ref List<UEBuildBinaryConfiguration> OutBuildBinaryConfigurations,
-		ref List<string> OutExtraModuleNames
-		)
-	{
-		OutExtraModuleNames.Add("ShaderPluginDemo");
-	}
+    public override void SetupBinaries(TargetInfo Target, ref List<UEBuildBinaryConfiguration> OutBuildBinaryConfigurations, ref List<string> OutExtraModuleNames)
+    {
+        OutExtraModuleNames.Add("ShaderPluginDemo");
+    }
 }
